refactor(provider): share history lookups
