+++ crates/storage/provider/src/providers/state/historical.rs
@@ -1,21 +1,15 @@
 use crate::{
-    AccountReader, BlockHashReader, ChangeSetReader, EitherReader, HashedPostStateProvider,
-    ProviderError, RocksDBProviderFactory, StateProvider, StateRootProvider,
+    AccountReader, BlockHashReader, ChangeSetReader, HashedPostStateProvider, ProviderError,
+    StateProvider, StateRootProvider,
 };
 use alloy_eips::merge::EPOCH_SLOTS;
 use alloy_primitives::{Address, BlockNumber, Bytes, StorageKey, StorageValue, B256};
-use reth_db_api::{
-    cursor::{DbCursorRO, DbDupCursorRO},
-    table::Table,
-    tables,
-    transaction::DbTx,
-    BlockNumberList,
-};
+use reth_db_api::{cursor::DbDupCursorRO, tables, transaction::DbTx};
 use reth_primitives_traits::{Account, Bytecode, NodePrimitives};
 use reth_storage_api::{
-    BlockNumReader, BytecodeReader, DBProvider, NodePrimitivesProvider, PruneCheckpointReader,
-    StageCheckpointReader, StateProofProvider, StorageChangeSetReader, StorageRootProvider,
-    StorageSettingsCache,
+    BlockNumReader, BytecodeReader, DBProvider, HistoryReader, NodePrimitivesProvider,
+    PruneCheckpointReader, StageCheckpointReader, StateProofProvider, StorageChangeSetReader,
+    StorageRootProvider, StorageSettingsCache,
 };
 use reth_storage_errors::provider::ProviderResult;
 use reth_storage_overlay::{OverlayManager, StateTrieOverlay};
@@ -32,6 +26,10 @@ use reth_trie::{
 use reth_trie_db::{DatabaseProof, DatabaseStateRoot, DatabaseStorageProof, DatabaseStorageRoot};
 
 use std::{fmt::Debug, sync::Arc};
+
+pub use reth_storage_api::{
+    compute_history_rank, history_info, needs_prev_shard_check, HistoryInfo,
+};
 
 type DbStateRoot<'a, TX, A> = StateRoot<
     reth_trie_db::DatabaseTrieCursorFactory<&'a TX, A>,
@@ -50,62 +48,6 @@ type DbProof<'a, TX, A> = Proof<
     reth_trie_db::DatabaseTrieCursorFactory<&'a TX, A>,
     reth_trie_db::DatabaseHashedCursorFactory<&'a TX>,
 >;
-
-/// Result of a history lookup for an account or storage slot.
-///
-/// Indicates where to find the historical value for a given key at a specific block.
-#[derive(Debug, Eq, PartialEq)]
-pub enum HistoryInfo {
-    /// The key is written to, but only after our block (not yet written at the target block). Or
-    /// it has never been written.
-    NotYetWritten,
-    /// The chunk contains an entry for a write after our block at the given block number.
-    /// The value should be looked up in the changeset at this block.
-    InChangeset(u64),
-    /// The chunk does not contain an entry for a write after our block. This can only
-    /// happen if this is the last chunk, so we need to look in the plain state.
-    InPlainState,
-    /// The key may have been written, but due to pruning we may not have changesets and
-    /// history, so we need to make a plain state lookup.
-    MaybeInPlainState,
-}
-
-impl HistoryInfo {
-    /// Determines where to find the historical value based on computed shard lookup results.
-    ///
-    /// This is a pure function shared by both MDBX and `RocksDB` backends.
-    ///
-    /// # Arguments
-    /// * `found_block` - The block number from the shard lookup
-    /// * `is_before_first_write` - True if the target block is before the first write to this key.
-    ///   This should be computed as: `rank == 0 && found_block != Some(block_number) &&
-    ///   !has_previous_shard` where `has_previous_shard` comes from a lazy `cursor.prev()` check.
-    /// * `lowest_available` - Lowest block where history is available (pruning boundary)
-    pub const fn from_lookup(
-        found_block: Option<u64>,
-        is_before_first_write: bool,
-        lowest_available: Option<BlockNumber>,
-    ) -> Self {
-        if is_before_first_write {
-            if let (Some(_), Some(block_number)) = (lowest_available, found_block) {
-                // The key may have been written, but due to pruning we may not have changesets
-                // and history, so we need to make a changeset lookup.
-                return Self::InChangeset(block_number)
-            }
-            // The key is written to, but only after our block.
-            return Self::NotYetWritten
-        }
-
-        if let Some(block_number) = found_block {
-            // The chunk contains an entry for a write after our block, return it.
-            Self::InChangeset(block_number)
-        } else {
-            // The chunk does not contain an entry for a write after our block. This can only
-            // happen if this is the last chunk and so we need to look in the plain state.
-            Self::InPlainState
-        }
-    }
-}
 
 /// State provider for a given block number which takes a tx reference.
 ///
@@ -170,29 +112,23 @@ where
         Self { provider, overlay_manager, block_number, lowest_available_blocks }
     }
 
-    /// Lookup an account in the `AccountsHistory` table using `EitherReader`.
+    /// Lookup an account in the `AccountsHistory` table.
     pub fn account_history_lookup(&self, address: Address) -> ProviderResult<HistoryInfo>
     where
-        Provider: StorageSettingsCache + RocksDBProviderFactory + NodePrimitivesProvider,
+        Provider: HistoryReader,
     {
         if !self.lowest_available_blocks.is_account_history_available(self.block_number) {
             return Err(ProviderError::StateAtBlockPruned(self.block_number))
         }
 
-        let visible_tip = self.provider.best_block_number()?;
-
-        self.provider.with_rocksdb_snapshot(|rocksdb_ref| {
-            let mut reader = EitherReader::new_accounts_history(self.provider, rocksdb_ref)?;
-            reader.account_history_info(
-                address,
-                self.block_number,
-                self.lowest_available_blocks.account_history_block_number,
-                visible_tip,
-            )
-        })
+        self.provider.account_history_info(
+            address,
+            self.block_number,
+            self.lowest_available_blocks.account_history_block_number,
+        )
     }
 
-    /// Lookup a storage key in the `StoragesHistory` table using `EitherReader`.
+    /// Lookup a storage key in the `StoragesHistory` table.
     ///
     /// `lookup_key` is always a plain (unhashed) storage key.
     pub fn storage_history_lookup(
@@ -201,24 +137,18 @@ where
         lookup_key: B256,
     ) -> ProviderResult<HistoryInfo>
     where
-        Provider: StorageSettingsCache + RocksDBProviderFactory + NodePrimitivesProvider,
+        Provider: HistoryReader,
     {
         if !self.lowest_available_blocks.is_storage_history_available(self.block_number) {
             return Err(ProviderError::StateAtBlockPruned(self.block_number))
         }
 
-        let visible_tip = self.provider.best_block_number()?;
-
-        self.provider.with_rocksdb_snapshot(|rocksdb_ref| {
-            let mut reader = EitherReader::new_storages_history(self.provider, rocksdb_ref)?;
-            reader.storage_history_info(
-                address,
-                lookup_key,
-                self.block_number,
-                self.lowest_available_blocks.storage_history_block_number,
-                visible_tip,
-            )
-        })
+        self.provider.storage_history_info(
+            address,
+            lookup_key,
+            self.block_number,
+            self.lowest_available_blocks.storage_history_block_number,
+        )
     }
 
     /// Resolves a storage value by looking up the given key in history, changesets, or
@@ -231,7 +161,7 @@ where
         lookup_key: B256,
     ) -> ProviderResult<Option<StorageValue>>
     where
-        Provider: StorageSettingsCache + RocksDBProviderFactory + NodePrimitivesProvider,
+        Provider: StorageSettingsCache + HistoryReader,
     {
         match self.storage_history_lookup(address, lookup_key)? {
             HistoryInfo::NotYetWritten => Ok(None),
@@ -344,7 +274,7 @@ where
         + ChangeSetReader
         + StorageChangeSetReader
         + StorageSettingsCache
-        + RocksDBProviderFactory
+        + HistoryReader
         + NodePrimitivesProvider<Primitives = N>,
     N: NodePrimitives,
 {
@@ -697,7 +627,7 @@ where
         + PruneCheckpointReader
         + StageCheckpointReader
         + StorageSettingsCache
-        + RocksDBProviderFactory
+        + HistoryReader
         + NodePrimitivesProvider<Primitives = N>,
     N: NodePrimitives,
 {
@@ -798,7 +728,7 @@ impl<
 }
 
 // Delegates all provider impls to [HistoricalStateProviderRef]
-reth_storage_api::macros::delegate_provider_impls!(HistoricalStateProvider<Provider> where [Provider: DBProvider + BlockNumReader + BlockHashReader + ChangeSetReader + StorageChangeSetReader + PruneCheckpointReader + StageCheckpointReader + StorageSettingsCache + RocksDBProviderFactory + NodePrimitivesProvider]);
+reth_storage_api::macros::delegate_provider_impls!(HistoricalStateProvider<Provider> where [Provider: DBProvider + BlockNumReader + BlockHashReader + ChangeSetReader + StorageChangeSetReader + PruneCheckpointReader + StageCheckpointReader + StorageSettingsCache + HistoryReader + NodePrimitivesProvider]);
 
 /// Lowest blocks at which different parts of the state are available.
 /// They may be [Some] if pruning is enabled.
@@ -828,97 +758,13 @@ impl LowestAvailableBlocks {
     }
 }
 
-/// Computes the rank and finds the next modification block in a history shard.
-///
-/// Given a `block_number`, this function returns:
-/// - `rank`: The number of entries strictly before `block_number` in the shard
-/// - `found_block`: The block number at position `rank` (i.e., the first block >= `block_number`
-///   where a modification occurred), or `None` if `rank` is out of bounds
-///
-/// The rank is adjusted when `block_number` exactly matches an entry in the shard,
-/// so that `found_block` always returns the modification at or after the target.
-///
-/// This logic is shared between MDBX cursor-based lookups and `RocksDB` iterator lookups.
-#[inline]
-pub fn compute_history_rank(
-    chunk: &reth_db_api::BlockNumberList,
-    block_number: BlockNumber,
-) -> (u64, Option<u64>) {
-    let mut rank = chunk.rank(block_number);
-    // `rank(block_number)` returns count of entries <= block_number.
-    // We want the first entry >= block_number, so if block_number is in the shard,
-    // we need to step back one position to point at it (not past it).
-    if rank.checked_sub(1).and_then(|r| chunk.select(r)) == Some(block_number) {
-        rank -= 1;
-    }
-    (rank, chunk.select(rank))
-}
-
-/// Checks if a previous shard lookup is needed to determine if we're before the first write.
-///
-/// Returns `true` when `rank == 0` (first entry in shard) and the found block doesn't match
-/// the target block number. In this case, we need to check if there's a previous shard.
-#[inline]
-pub fn needs_prev_shard_check(
-    rank: u64,
-    found_block: Option<u64>,
-    block_number: BlockNumber,
-) -> bool {
-    rank == 0 && found_block != Some(block_number)
-}
-
-/// Generic history lookup for sharded history tables.
-///
-/// Seeks to the shard containing `block_number`, verifies the key via `key_filter`,
-/// and checks previous shard to detect if we're before the first write.
-pub fn history_info<T, K, C>(
-    cursor: &mut C,
-    key: K,
-    block_number: BlockNumber,
-    key_filter: impl Fn(&K) -> bool,
-    lowest_available_block_number: Option<BlockNumber>,
-) -> ProviderResult<HistoryInfo>
-where
-    T: Table<Key = K, Value = BlockNumberList>,
-    C: DbCursorRO<T>,
-{
-    // Lookup the history chunk in the history index. If the key does not appear in the
-    // index, the first chunk for the next key will be returned so we filter out chunks that
-    // have a different key.
-    if let Some(chunk) = cursor.seek(key)?.filter(|(k, _)| key_filter(k)).map(|x| x.1) {
-        let (rank, found_block) = compute_history_rank(&chunk, block_number);
-
-        // If our block is before the first entry in the index chunk and this first entry
-        // doesn't equal to our block, it might be before the first write ever. To check, we
-        // look at the previous entry and check if the key is the same.
-        // This check is worth it, the `cursor.prev()` check is rarely triggered (the if will
-        // short-circuit) and when it passes we save a full seek into the changeset/plain state
-        // table.
-        let is_before_first_write = needs_prev_shard_check(rank, found_block, block_number) &&
-            !cursor.prev()?.is_some_and(|(k, _)| key_filter(&k));
-
-        Ok(HistoryInfo::from_lookup(
-            found_block,
-            is_before_first_write,
-            lowest_available_block_number,
-        ))
-    } else if lowest_available_block_number.is_some() {
-        // The key may have been written, but due to pruning we may not have changesets and
-        // history, so we need to make a plain state lookup.
-        Ok(HistoryInfo::MaybeInPlainState)
-    } else {
-        // The key has not been written to at all.
-        Ok(HistoryInfo::NotYetWritten)
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::needs_prev_shard_check;
     use crate::{
         providers::state::historical::{HistoryInfo, LowestAvailableBlocks},
         test_utils::create_test_provider_factory,
-        AccountReader, HistoricalStateProvider, HistoricalStateProviderRef, RocksDBProviderFactory,
+        AccountReader, HistoricalStateProvider, HistoricalStateProviderRef, HistoryReader,
         StateProvider,
     };
     use alloy_primitives::{address, b256, Address, B256, U256};
@@ -953,7 +799,7 @@ mod tests {
             + PruneCheckpointReader
             + StageCheckpointReader
             + StorageSettingsCache
-            + RocksDBProviderFactory
+            + HistoryReader
             + NodePrimitivesProvider,
     >() {
         assert_state_provider::<HistoricalStateProvider<T>>();
